--- docs/.vitepress/versions.json
@@ -1,11 +1,0 @@
-{
-  "versions": [
-    { "text": "main", "link": "https://docs.px4.io/main/en/" },
-    { "text": "v1.17 (stable)", "link": "https://docs.px4.io/v1.17/en/" },
-    { "text": "v1.16", "link": "https://docs.px4.io/v1.16/en/" },
-    { "text": "v1.15", "link": "https://docs.px4.io/v1.15/en/" },
-    { "text": "v1.14", "link": "https://docs.px4.io/v1.14/en/" },
-    { "text": "v1.13", "link": "https://docs.px4.io/v1.13/en/" },
-    { "text": "v1.12", "link": "https://docs.px4.io/v1.12/en/" }
-  ]
-}
